TCP --receive-tcp and --send-tcp functionality
Major logic change, the vectors holding the output channels are now wrapped in an Arc/Mutex so that we can thread-safe write values in to it once a socket turns up. This way we can non-block the startup routine.

Test suite updated. The test for this is super slow but I couldn't be bothered to write better python to work around it.